feat: 재고 증감 기능 구현
